+++ Engine/Logic/Physics/JoltUtils.cpp
@@ -1,0 +1,54 @@
+#include "JoltUtils.h"
+#include <cstdarg>
+
+namespace ToolEngine
+{
+	BPLayerInterfaceImpl::BPLayerInterfaceImpl()
+	{
+		m_object_to_broad_phase[Layers::NON_MOVING] = BroadPhaseLayers::NON_MOVING;
+		m_object_to_broad_phase[Layers::MOVING] = BroadPhaseLayers::MOVING;
+	}
+	unsigned int BPLayerInterfaceImpl::GetNumBroadPhaseLayers() const
+	{
+		return BroadPhaseLayers::NUM_LAYERS;
+	}
+	JPH::BroadPhaseLayer BPLayerInterfaceImpl::GetBroadPhaseLayer(JPH::ObjectLayer inLayer) const
+	{
+		return m_object_to_broad_phase[inLayer];
+	}
+	const char* BPLayerInterfaceImpl::GetBroadPhaseLayerName(JPH::BroadPhaseLayer inLayer) const
+	{
+		switch ((JPH::BroadPhaseLayer::Type)inLayer)
+		{
+		case (JPH::BroadPhaseLayer::Type)BroadPhaseLayers::NON_MOVING:	return "NON_MOVING";
+		case (JPH::BroadPhaseLayer::Type)BroadPhaseLayers::MOVING:		return "MOVING";
+		default: JPH_ASSERT(false); return "INVALID";
+		}
+	}
+	bool ObjectVsBroadPhaseLayerFilterImpl::ShouldCollide(JPH::ObjectLayer inLayer1, JPH::BroadPhaseLayer inLayer2) const
+	{
+		switch (inLayer1)
+		{
+		case Layers::NON_MOVING:
+			return inLayer2 == BroadPhaseLayers::MOVING;
+		case Layers::MOVING:
+			return true;
+		default:
+			JPH_ASSERT(false);
+			return false;
+		}
+	}
+	bool ObjectLayerPairFilterImpl::ShouldCollide(JPH::ObjectLayer inObject1, JPH::ObjectLayer inObject2) const
+	{
+		switch (inObject1)
+		{
+		case Layers::NON_MOVING:
+			return inObject2 == Layers::MOVING; // Non moving only collides with moving
+		case Layers::MOVING:
+			return true; // Moving collides with everything
+		default:
+			JPH_ASSERT(false);
+			return false;
+		}
+	}
+}
