+++ Engine/Logic/Physics/JoltUtils.h
@@ -1,0 +1,54 @@
+#pragma once
+
+#include "Marco.h"
+#include <Jolt/Jolt.h>
+#include <Jolt/RegisterTypes.h>
+#include <Jolt/Core/Factory.h>
+#include <Jolt/Core/TempAllocator.h>
+#include <Jolt/Core/JobSystemThreadPool.h>
+#include <Jolt/Physics/PhysicsSettings.h>
+#include <Jolt/Physics/PhysicsSystem.h>
+#include <Jolt/Physics/Collision/Shape/BoxShape.h>
+#include <Jolt/Physics/Collision/Shape/SphereShape.h>
+#include <Jolt/Physics/Body/BodyCreationSettings.h>
+#include <Jolt/Physics/Body/BodyActivationListener.h>
+
+namespace ToolEngine
+{
+	namespace Layers
+	{
+		static constexpr JPH::ObjectLayer NON_MOVING = 0;
+		static constexpr JPH::ObjectLayer MOVING = 1;
+		static constexpr JPH::ObjectLayer NUM_LAYERS = 2;
+	};
+	namespace BroadPhaseLayers
+	{
+		static constexpr JPH::BroadPhaseLayer NON_MOVING(0);
+		static constexpr JPH::BroadPhaseLayer MOVING(1);
+		static constexpr unsigned int NUM_LAYERS(2);
+	};
+
+	// Broadphase use Axis test
+	class BPLayerInterfaceImpl final : public JPH::BroadPhaseLayerInterface
+	{
+	public:
+		BPLayerInterfaceImpl();
+		virtual unsigned int GetNumBroadPhaseLayers() const override;
+		virtual JPH::BroadPhaseLayer GetBroadPhaseLayer(JPH::ObjectLayer inLayer) const override;
+		virtual const char* GetBroadPhaseLayerName(JPH::BroadPhaseLayer inLayer) const override;
+	private:
+		JPH::BroadPhaseLayer m_object_to_broad_phase[Layers::NUM_LAYERS];
+	};
+
+	class ObjectVsBroadPhaseLayerFilterImpl : public JPH::ObjectVsBroadPhaseLayerFilter
+	{
+	public:
+		virtual bool ShouldCollide(JPH::ObjectLayer inLayer1, JPH::BroadPhaseLayer inLayer2) const override;
+	};
+
+	class ObjectLayerPairFilterImpl : public JPH::ObjectLayerPairFilter
+	{
+	public:
+		virtual bool ShouldCollide(JPH::ObjectLayer inObject1, JPH::ObjectLayer inObject2) const override;
+	};
+}
