+++ Engine/Logic/Physics/PhysicsManager.h
@@ -13,46 +13,10 @@
 #include <Jolt/Physics/Body/BodyCreationSettings.h>
 #include <Jolt/Physics/Body/BodyActivationListener.h>
 #include "Geometry/RenderScene.h"
+#include "JoltUtils.h"
 
 namespace ToolEngine
 {
-	namespace Layers
-	{
-		static constexpr JPH::ObjectLayer NON_MOVING = 0;
-		static constexpr JPH::ObjectLayer MOVING = 1;
-		static constexpr JPH::ObjectLayer NUM_LAYERS = 2;
-	};
-	namespace BroadPhaseLayers
-	{
-		static constexpr JPH::BroadPhaseLayer NON_MOVING(0);
-		static constexpr JPH::BroadPhaseLayer MOVING(1);
-		static constexpr unsigned int NUM_LAYERS(2);
-	};
-
-	// Broadphase use Axis test
-	class BPLayerInterfaceImpl final : public JPH::BroadPhaseLayerInterface
-	{
-	public:
-		BPLayerInterfaceImpl();
-		virtual unsigned int GetNumBroadPhaseLayers() const override;
-		virtual JPH::BroadPhaseLayer GetBroadPhaseLayer(JPH::ObjectLayer inLayer) const override;
-		virtual const char* GetBroadPhaseLayerName(JPH::BroadPhaseLayer inLayer) const override;
-	private:
-		JPH::BroadPhaseLayer m_object_to_broad_phase[Layers::NUM_LAYERS];
-	};
-
-	class ObjectVsBroadPhaseLayerFilterImpl : public JPH::ObjectVsBroadPhaseLayerFilter
-	{
-	public:
-		virtual bool ShouldCollide(JPH::ObjectLayer inLayer1, JPH::BroadPhaseLayer inLayer2) const override;
-	};
-
-	class ObjectLayerPairFilterImpl : public JPH::ObjectLayerPairFilter
-	{
-	public:
-		virtual bool ShouldCollide(JPH::ObjectLayer inObject1, JPH::ObjectLayer inObject2) const override;
-	};
-
 	class PhysicsManager
 	{
 	public:
